+++ index.d.ts
@@ -1,9 +1,12 @@
+import { PubSubEngine } from "graphql-subscriptions";
+import { ServerOptions } from "graphql-ws";
+
 declare module "moleculer-apollo-server" {
 	import { ServiceSchema, Context } from "moleculer";
 	import { Config } from "apollo-server-core";
 	import { OptionsUrlencoded } from "body-parser";
 	import { SchemaDirectiveVisitor, IResolvers } from "graphql-tools";
-
+	import { RenderPageOptions } from "@apollographql/graphql-playground-html";
 	export {
 		GraphQLExtension,
 		gql,
@@ -17,6 +20,7 @@ declare module "moleculer-apollo-server" {
 		defaultPlaygroundOptions,
 	} from "apollo-server-core";
 
+	
 	export { GraphQLUpload } from "graphql-upload";
 
 	export * from "graphql-tools";
@@ -86,6 +90,18 @@ declare module "moleculer-apollo-server" {
 		onBeforeCall?: (ctx: Context, route: any, req: any, res: any) => Promise<any>;
 		onAfterCall?: (ctx: Context, route: any, req: any, res: any, data: any) => Promise<any>;
 	}
+   // Promise<GraphQLExecutionContextValue> | GraphQLExecutionContextValue)
+
+	export interface SubscriptionsConfig  {
+		onConnect?:(ctx: Context) => Promise<Record<string, unknown> | boolean | void> | Record<string, unknown> | boolean | void;
+		context?:(ctx:Context) => Promise<any>;
+		createPubSub?:() => typeof PubSubEngine;
+	}
+
+	export interface ServerMixinOptions extends Config {
+		playgroundOptions?:RenderPageOptions;
+		subscriptions?:SubscriptionsConfig;
+	}
 
 	export interface ApolloServiceOptions {
 		typeDefs?: string | string[];
@@ -94,7 +110,7 @@ declare module "moleculer-apollo-server" {
 			[name: string]: typeof SchemaDirectiveVisitor;
 		};
 		routeOptions?: ServiceRouteOptions;
-		serverOptions?: Config;
+		serverOptions?: ServerMixinOptions;
 		checkActionVisibility?: boolean;
 		autoUpdateSchema?: boolean;
 	}