+++ src/ApolloServer.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const { ApolloServerBase } = require("apollo-server-core");
-// const { processRequest } = require("graphql-upload");
+const { processRequest } = require("graphql-upload");
 const { renderPlaygroundPage } = require("@apollographql/graphql-playground-html");
 const accept = require("@hapi/accept");
 const moleculerApollo = require("./moleculerApollo");
@@ -31,7 +31,7 @@ class ApolloServer extends ApolloServerBase {
 
 	// Prepares and returns an async function that can be used to handle
 	// GraphQL requests.
-	createHandler({ path, disableHealthCheck, onHealthCheck } = {}) {
+	createHandler({ path, disableHealthCheck, onHealthCheck, playgroundOptions } = {}) {
 		// const promiseWillStart = this.willStart();
 
 		return async (req, res) => {
@@ -44,7 +44,7 @@ class ApolloServer extends ApolloServerBase {
 			if (this.uploadsConfig) {
 				const contentType = req.headers["content-type"];
 				if (contentType && contentType.startsWith("multipart/form-data")) {
-					// req.filePayload = await processRequest(req, res, this.uploadsConfig);
+					req.filePayload = await processRequest(req, res, this.uploadsConfig);
 				}
 			}
 
@@ -56,7 +56,7 @@ class ApolloServer extends ApolloServerBase {
 			// If the `playgroundOptions` are set, register a `graphql-playground` instance
 			// (not available in production) that is then used to handle all
 			// incoming GraphQL requests.
-			if (/*this.playgroundOptions &&*/ req.method === "GET") {
+			if (playgroundOptions && req.method === "GET") {
 				const { mediaTypes } = accept.parseAll(req.headers);
 				const prefersHTML =
 					mediaTypes.find(x => x === "text/html" || x === "application/json") ===
@@ -66,9 +66,9 @@ class ApolloServer extends ApolloServerBase {
 					const middlewareOptions = Object.assign(
 						{
 							endpoint: this.graphqlPath,
-							subscriptionEndpoint: this.subscriptionsPath,
+							subscriptionEndpoint: this.subscriptionsPath || this.graphqlPath,
 						},
-						this.playgroundOptions
+						playgroundOptions
 					);
 					return send(
 						req,